+++ tests/testthat/test-list-modify-update.R
@@ -1,4 +1,4 @@
-context("list_modify")
+context("list-modify")
 
 # list_modify -------------------------------------------------------------
 
@@ -11,14 +11,27 @@ test_that("named lists have values replaced by name", {
 test_that("unnamed lists are replaced by position", {
   expect_equal(list_modify(list(3), 1, 2), list(1, 2))
   expect_equal(list_modify(list(1, 2, 3), 4), list(4, 2, 3))
-  expect_equal(list_modify(list(1, 2, 3), NULL, NULL), list(3))
 })
 
-test_that("error if one named and the other is not", {
+test_that("can remove elements with `zap()`", {
+  expect_equal(list_modify(list(1, 2, 3), zap(), zap()), list(3))
+  expect_equal(list_modify(list(a = 1, b = 2, c = 3), b = zap(), a = zap()), list(c = 3))
+})
+
+test_that("error if inputs are not all named or unnamed", {
   expect_error(
-    list_modify(list(a = 1), 2),
-    "must be either both named or both unnamed"
+    list_modify(list(a = 1), 2, a = 2),
+    "must be either all named, or all unnamed"
   )
+})
+
+test_that("can update unnamed lists with named inputs", {
+  expect_identical(list_modify(list(1), a = 2), list(1, a = 2))
+})
+
+test_that("can update named lists with unnamed inputs", {
+  expect_identical(list_modify(list(a = 1, b = 2), 2), list(a = 2, b = 2))
+  expect_identical(list_modify(list(a = 1, b = 2), 2, 3, 4), list(a = 2, b = 3, 4))
 })
 
 test_that("lists are replaced recursively", {
@@ -82,4 +95,17 @@ test_that("can modify element called x", {
 test_that("quosures and formulas are evaluated", {
   expect_identical(update_list(list(x = 1), y = quo(x + 1)), list(x = 1, y = 2))
   expect_identical(update_list(list(x = 1), y = ~x + 1), list(x = 1, y = 2))
+})
+
+
+# Life cycle --------------------------------------------------------------
+
+test_that("removing elements with `NULL` is deprecated", {
+  scoped_lifecycle_warnings()
+  expect_warning(list_modify(list(1, 2, 3), NULL, NULL), list(3), "deprecated")
+})
+
+test_that("can still remove elements with `NULL`", {
+  scoped_options(lifecycle_disable_warnings = TRUE)
+  expect_equal(list_modify(list(1, 2, 3), NULL, NULL), list(3))
 })